% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{arguments_to_md_table}
\alias{arguments_to_md_table}
\title{Convert Rd arguments section to Markdown table}
\usage{
arguments_to_md_table(rd_obj)
}
\arguments{
\item{rd_obj}{Rd object}
}
\value{
Markdown table string or NULL if no arguments
}
\description{
Convert Rd arguments section to Markdown table
}
\keyword{internal}
