% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{build_frontmatter}
\alias{build_frontmatter}
\title{Build YAML frontmatter string}
\usage{
build_frontmatter(frontmatter_data)
}
\arguments{
\item{frontmatter_data}{List from extract_frontmatter_data}
}
\value{
Character string with YAML frontmatter
}
\description{
Build YAML frontmatter string
}
\keyword{internal}
