% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_example_content.R
\name{capture_example_output}
\alias{capture_example_output}
\title{Runs example code and capture output to a file}
\usage{
capture_example_output(pkg_name, artifact_output_dir, text_output_dir)
}
\arguments{
\item{pkg_name}{name of package to collect Rd objects for}

\item{artifact_output_dir}{path to directory to save ggplots and gt tables}

\item{text_output_dir}{path to directory to save text output files}
}
\description{
Runs example code and capture output to a file
}
\examples{
\dontrun{
capture_example_output("pkg", "pkg-docs/public")
}
}
