% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{check_pandoc}
\alias{check_pandoc}
\title{Check if pandoc is available}
\usage{
check_pandoc()
}
\value{
Logical, TRUE if pandoc is available
}
\description{
Check if pandoc is available
}
\keyword{internal}
