% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_sections.R
\name{code_section_to_md}
\alias{code_section_to_md}
\title{Convert Rd code section (usage/examples) to Markdown code block}
\usage{
code_section_to_md(rd_obj, section)
}
\arguments{
\item{rd_obj}{Rd object}

\item{section}{Section name ("usage" or "examples")}
}
\value{
Markdown code block string or NULL
}
\description{
Convert Rd code section (usage/examples) to Markdown code block
}
\keyword{internal}
