% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown_converter.R
\name{convert_articles_structure}
\alias{convert_articles_structure}
\title{Convert pkgdown articles structure to starlightr format}
\usage{
convert_articles_structure(articles)
}
\arguments{
\item{articles}{Pkgdown articles section}
}
\value{
Starlightr articles structure
}
\description{
Convert pkgdown articles structure to starlightr format
}
