% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown_converter.R
\name{convert_reference_structure}
\alias{convert_reference_structure}
\title{Convert pkgdown reference structure to starlightr format}
\usage{
convert_reference_structure(reference)
}
\arguments{
\item{reference}{Pkgdown reference section}
}
\value{
Starlightr reference structure
}
\description{
Convert pkgdown reference structure to starlightr format
}
