% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{copy_assets}
\alias{copy_assets}
\title{Copy assets from package to site}
\usage{
copy_assets(pkg_path, output_path, config)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Copy assets from package to site
}
