% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{copy_branding_assets}
\alias{copy_branding_assets}
\title{Copy logo and favicon to site}
\usage{
copy_branding_assets(pkg_path, output_path, config)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Copy logo and favicon to site
}
\keyword{internal}
