% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_template}
\alias{copy_template}
\title{Copy a static template from inst/templates/ to a destination}
\usage{
copy_template(name, dest)
}
\arguments{
\item{name}{Template filename (e.g. "gitignore")}

\item{dest}{Destination file path}
}
\description{
Copy a static template from inst/templates/ to a destination
}
\keyword{internal}
