% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{create_default_config}
\alias{create_default_config}
\title{Create default _starlightr.yaml configuration}
\usage{
create_default_config(config_path, pkg_path)
}
\arguments{
\item{config_path}{Path where to create the config file}

\item{pkg_path}{Path to package directory}
}
\description{
Create default _starlightr.yaml configuration
}
