% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{create_directory_indexes}
\alias{create_directory_indexes}
\title{Create directory index pages for articles and reference}
\usage{
create_directory_indexes(pkg_path, output_path, config)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Create directory index pages for articles and reference
}
\keyword{internal}
