% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_starlightr.R
\name{create_sample_article}
\alias{create_sample_article}
\title{Create a sample article in vignettes/}
\usage{
create_sample_article(articles_dir)
}
\arguments{
\item{articles_dir}{Path to vignettes directory}
}
\description{
Create a sample article in vignettes/
}
\keyword{internal}
