% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{default_config}
\alias{default_config}
\title{Get default configuration}
\usage{
default_config()
}
\value{
List with default configuration values
}
\description{
Get default configuration
}
\keyword{internal}
