% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_dir}
\alias{ensure_dir}
\title{Create a directory, aborting on failure}
\usage{
ensure_dir(path)
}
\arguments{
\item{path}{Directory path to create}
}
\value{
Invisibly returns the path
}
\description{
Create a directory, aborting on failure
}
\keyword{internal}
