% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{escape_angle_brackets}
\alias{escape_angle_brackets}
\title{Escape angle brackets outside code blocks for MDX}
\usage{
escape_angle_brackets(md)
}
\arguments{
\item{md}{Markdown string}
}
\value{
Markdown with < escaped outside code blocks
}
\description{
Escape angle brackets outside code blocks for MDX
}
\keyword{internal}
