% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape special regex characters in a pattern}
\usage{
escape_regex(pattern)
}
\arguments{
\item{pattern}{String to escape}
}
\value{
Escaped string
}
\description{
Escape special regex characters in a pattern
}
\keyword{internal}
