% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{evaluate_sexpr}
\alias{evaluate_sexpr}
\title{Evaluate unevaluated Sexpr expressions in HTML}
\usage{
evaluate_sexpr(html)
}
\arguments{
\item{html}{HTML string potentially containing unevaluated Sexpr}
}
\value{
HTML with Sexpr expressions evaluated
}
\description{
Rd2HTML doesn't evaluate stage=render Sexprs when converting statically.
This function finds and evaluates them manually.
}
\keyword{internal}
