% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{expand_glob_patterns}
\alias{expand_glob_patterns}
\title{Expand glob patterns to matching files}
\usage{
expand_glob_patterns(patterns, available_files)
}
\arguments{
\item{patterns}{Vector of patterns (may include * wildcards)}

\item{available_files}{Vector of available file names}
}
\value{
Vector of matched file names
}
\description{
Expand glob patterns to matching files
}
\keyword{internal}
