% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{expand_pkgdown_selectors}
\alias{expand_pkgdown_selectors}
\title{Expand pkgdown selector functions to actual function names}
\usage{
expand_pkgdown_selectors(contents, output_path = NULL)
}
\arguments{
\item{contents}{Vector of function names and selectors}

\item{output_path}{Path to output directory to find available functions}
}
\value{
Vector with selectors expanded to actual function names
}
\description{
Expand pkgdown selector functions to actual function names
}
