% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_links.R
\name{extract_dependency_packages}
\alias{extract_dependency_packages}
\title{Extract dependency package names from a DESCRIPTION file}
\usage{
extract_dependency_packages(package = ".")
}
\arguments{
\item{package}{Path to package root containing DESCRIPTION}
}
\value{
Character vector of package names
}
\description{
Extract dependency package names from a DESCRIPTION file
}
\keyword{internal}
