% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_example_output.R
\name{extract_examples_code}
\alias{extract_examples_code}
\title{Extract raw code from an Rd examples section}
\usage{
extract_examples_code(examples_section)
}
\arguments{
\item{examples_section}{The examples section from an Rd object}
}
\value{
Character string of example code
}
\description{
Extract raw code from an Rd examples section
}
\keyword{internal}
