% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{extract_frontmatter_data}
\alias{extract_frontmatter_data}
\title{Extract frontmatter data from Rd object}
\usage{
extract_frontmatter_data(rd_obj)
}
\arguments{
\item{rd_obj}{Rd object}
}
\value{
List with title, description, and aliases
}
\description{
Extract frontmatter data from Rd object
}
\keyword{internal}
