% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown_converter.R
\name{extract_github_url}
\alias{extract_github_url}
\title{Extract GitHub URL from various pkgdown locations}
\usage{
extract_github_url(pkgdown)
}
\arguments{
\item{pkgdown}{Parsed pkgdown YAML}
}
\value{
GitHub URL or NULL
}
\description{
Extract GitHub URL from various pkgdown locations
}
