% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_rd_content}
\alias{extract_rd_content}
\title{Extracts all sections from an Rd object dynamically}
\usage{
extract_rd_content(rd_obj, include_common_only = FALSE)
}
\arguments{
\item{rd_obj}{rd object from tools::Rd_db}

\item{include_common_only}{logical, if TRUE only extracts common sections,
if FALSE extracts all available sections}
}
\value{
list of sections for the rd_obj
}
\description{
Extracts all sections from an Rd object dynamically
}
\examples{
\dontrun{
rd_files <- tools::Rd_db("starlightr")

# Extract all sections
extract_rd_content(rd_files[["extract_section.Rd"]])

# Extract only common sections
extract_rd_content(rd_files[["extract_section.Rd"]], include_common_only = TRUE)
}
}
