% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_rd_content_common}
\alias{extract_rd_content_common}
\title{Wrapper for backward compatibility - extracts only common sections}
\usage{
extract_rd_content_common(rd_obj)
}
\arguments{
\item{rd_obj}{rd object from tools::Rd_db}
}
\value{
list of common sections for the rd_obj
}
\description{
Wrapper for backward compatibility - extracts only common sections
}
