% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_section}
\alias{extract_section}
\title{Pulls out text from rd_obj if Rd_tag attribute matches supplied tag}
\usage{
extract_section(rd_obj, tag)
}
\arguments{
\item{rd_obj}{rd object from tools::Rd_db}

\item{tag}{Rd section like name, description, arguments...}
}
\value{
a string of the tagged section
}
\description{
Pulls out text from rd_obj if Rd_tag attribute matches supplied tag
}
\examples{
\dontrun{
rd_files <- tools::Rd_db("starlightr")

extract_section(rd_files[["extract_section.Rd"]], "description")
}
}
