% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown_converter.R
\name{extract_site_info}
\alias{extract_site_info}
\title{Extract site information from pkgdown config}
\usage{
extract_site_info(pkgdown, pkg_path)
}
\arguments{
\item{pkgdown}{Parsed pkgdown YAML}

\item{pkg_path}{Path to package directory}
}
\value{
List with site and home information
}
\description{
Extract site information from pkgdown config
}
