% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{find_link_entries}
\alias{find_link_entries}
\title{Recursively find link fields in config for audits}
\usage{
find_link_entries(config)
}
\arguments{
\item{config}{Parsed YAML configuration list}
}
\value{
List of link entries (link + context)
}
\description{
Recursively find link fields in config for audits
}
\keyword{internal}
