% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_readme}
\alias{find_readme}
\title{Find README.Rmd in package root or inst/}
\usage{
find_readme(pkg_path)
}
\arguments{
\item{pkg_path}{Path to package directory}
}
\value{
README path if found, otherwise NULL
}
\description{
Find README.Rmd in package root or inst/
}
\keyword{internal}
