% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_postprocess.R
\name{fix_figure_paths}
\alias{fix_figure_paths}
\title{Fix figure paths for Starlight}
\usage{
fix_figure_paths(md)
}
\arguments{
\item{md}{Markdown string}
}
\value{
Markdown with fixed figure paths
}
\description{
Fix figure paths for Starlight
}
\keyword{internal}
