% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{fix_heading_levels}
\alias{fix_heading_levels}
\title{Fix heading levels in Markdown}
\usage{
fix_heading_levels(md)
}
\arguments{
\item{md}{Markdown string}
}
\value{
Markdown string with adjusted headings
}
\description{
Rd2HTML produces h3 headings; we want h2 for main sections.
}
\keyword{internal}
