% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_postprocess.R
\name{fix_lifecycle_badges}
\alias{fix_lifecycle_badges}
\title{Fix lifecycle badge paths to use CDN}
\usage{
fix_lifecycle_badges(md)
}
\arguments{
\item{md}{Markdown string}
}
\value{
Markdown with lifecycle badges pointing to CDN
}
\description{
Fix lifecycle badge paths to use CDN
}
\keyword{internal}
