% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_md.R
\name{format_arguments_section}
\alias{format_arguments_section}
\title{Format arguments section as a table}
\usage{
format_arguments_section(arguments_content)
}
\arguments{
\item{arguments_content}{raw arguments content from Rd file}
}
\value{
formatted markdown string
}
\description{
Format arguments section as a table
}
