% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_md.R
\name{format_md}
\alias{format_md}
\title{Formats the rd file contents to markdown}
\usage{
format_md(
  content,
  sections = c("description", "usage", "arguments", "details", "value", "examples",
    "references", "note", "author", "source", "format", "section", "subsection"),
  code_sections = c("usage", "examples"),
  skip_sections = c("name", "title", "seealso")
)
}
\arguments{
\item{content}{rd_file contents from extract_rd_content}

\item{sections}{character vector of sections to include, in order}

\item{code_sections}{character vector of sections to format as code blocks}

\item{skip_sections}{character vector of sections to skip entirely}
}
\value{
markdown formatted string of content
}
\description{
Formats the rd file contents to markdown
}
\examples{
\dontrun{
rd_files <- extract_package_rd_content("rdstarlight")
format_md(rd_files[["function_name.Rd"]])
}
}
