% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{generate_astro_config}
\alias{generate_astro_config}
\title{Generate astro.config.mjs file}
\usage{
generate_astro_config(output_path, config)
}
\arguments{
\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Generate astro.config.mjs file
}
