% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_gitignore}
\alias{generate_gitignore}
\title{Generate .gitignore file for Starlight site}
\usage{
generate_gitignore(output_path)
}
\arguments{
\item{output_path}{Path to output directory}
}
\description{
Generate .gitignore file for Starlight site
}
\keyword{internal}
