% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_package_json}
\alias{generate_package_json}
\title{Generate package.json file}
\usage{
generate_package_json(output_path, config)
}
\arguments{
\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Generate package.json file
}
\keyword{internal}
