% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{generate_starlight_config}
\alias{generate_starlight_config}
\title{Generate Starlight configuration files}
\usage{
generate_starlight_config(output_path, config, pkg_path)
}
\arguments{
\item{output_path}{Path to output directory}

\item{config}{Configuration list}

\item{pkg_path}{Path to package directory}
}
\description{
Generate Starlight configuration files
}
