% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{generate_version_select_component}
\alias{generate_version_select_component}
\title{Generate VersionSelect.astro component}
\usage{
generate_version_select_component(output_path, config)
}
\arguments{
\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Generate VersionSelect.astro component
}
\keyword{internal}
