% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{generate_versions_ts}
\alias{generate_versions_ts}
\title{Generate versions.ts data file}
\usage{
generate_versions_ts(output_path, config)
}
\arguments{
\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Generate versions.ts data file
}
\keyword{internal}
