% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{get_available_ref_files}
\alias{get_available_ref_files}
\title{Get available reference files}
\usage{
get_available_ref_files(output_path)
}
\arguments{
\item{output_path}{Output path}
}
\value{
Character vector of file names without extension
}
\description{
Get available reference files
}
\keyword{internal}
