% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{get_example_outputs}
\alias{get_example_outputs}
\title{Get example outputs for embedding}
\usage{
get_example_outputs(func_name, output_path)
}
\arguments{
\item{func_name}{Function name}

\item{output_path}{Path to site output directory}
}
\value{
Markdown string with embedded outputs or NULL
}
\description{
Get example outputs for embedding
}
\keyword{internal}
