% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{get_package_name}
\alias{get_package_name}
\title{Get package name from DESCRIPTION file}
\usage{
get_package_name(pkg_path)
}
\arguments{
\item{pkg_path}{Path to package directory}
}
\value{
Package name as string
}
\description{
Get package name from DESCRIPTION file
}
