% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{get_rd_aliases}
\alias{get_rd_aliases}
\title{Get all aliases from an Rd object}
\usage{
get_rd_aliases(rd_obj)
}
\arguments{
\item{rd_obj}{Rd object from tools::Rd_db}
}
\value{
Character vector of aliases
}
\description{
Get all aliases from an Rd object
}
\keyword{internal}
