% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_rd_tags}
\alias{get_rd_tags}
\title{Get all available Rd tags from an rd object}
\usage{
get_rd_tags(rd_obj)
}
\arguments{
\item{rd_obj}{rd object from tools::Rd_db}
}
\value{
character vector of all Rd_tag values found in the object
}
\description{
Get all available Rd tags from an rd object
}
