% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{html_to_md}
\alias{html_to_md}
\title{Convert HTML string to Markdown using pandoc}
\usage{
html_to_md(html, wrap = FALSE)
}
\arguments{
\item{html}{Character string containing HTML}

\item{wrap}{Logical, whether to wrap long lines (default FALSE)}
}
\value{
Character string containing Markdown
}
\description{
Convert HTML string to Markdown using pandoc
}
\keyword{internal}
