% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{is_absolute_path}
\alias{is_absolute_path}
\title{Check if a path is absolute}
\usage{
is_absolute_path(path)
}
\arguments{
\item{path}{Character string representing a file path}
}
\value{
Logical indicating if the path is absolute
}
\description{
Check if a path is absolute
}
