% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{is_internal_link}
\alias{is_internal_link}
\title{Determine if a link is internal}
\usage{
is_internal_link(link)
}
\arguments{
\item{link}{Link string}
}
\value{
Logical indicating if link is internal
}
\description{
Determine if a link is internal
}
\keyword{internal}
