% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{is_pkgdown_selector}
\alias{is_pkgdown_selector}
\title{Check if a string is a pkgdown selector function}
\usage{
is_pkgdown_selector(content)
}
\arguments{
\item{content}{String to check}
}
\value{
Logical indicating if it's a selector
}
\description{
Check if a string is a pkgdown selector function
}
