% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{link_needs_trailing_slash}
\alias{link_needs_trailing_slash}
\title{Check if an internal link needs a trailing slash}
\usage{
link_needs_trailing_slash(link)
}
\arguments{
\item{link}{Link string}
}
\value{
Logical indicating if link should end with '/'
}
\description{
Check if an internal link needs a trailing slash
}
\keyword{internal}
