% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_md.R
\name{looks_like_list_item}
\alias{looks_like_list_item}
\title{Check if a line looks like it could be a list item}
\usage{
looks_like_list_item(line)
}
\arguments{
\item{line}{text line to check}
}
\value{
logical
}
\description{
Check if a line looks like it could be a list item
}
