% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{make_sidebar_item}
\alias{make_sidebar_item}
\title{Create a single sidebar item JSON string}
\usage{
make_sidebar_item(label, slug)
}
\arguments{
\item{label}{Display label}

\item{slug}{URL slug}
}
\value{
JSON string for sidebar item
}
\description{
Create a single sidebar item JSON string
}
\keyword{internal}
