% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{merge_config}
\alias{merge_config}
\title{Merge user config with defaults}
\usage{
merge_config(user_config, default_config)
}
\arguments{
\item{user_config}{User-provided configuration}

\item{default_config}{Default configuration}
}
\value{
Merged configuration list
}
\description{
Merge user config with defaults
}
