% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{merge_package_deps}
\alias{merge_package_deps}
\title{Merge katex and user dependencies into package.json}
\usage{
merge_package_deps(package_json_path, config)
}
\arguments{
\item{package_json_path}{Path to the package.json file}

\item{config}{Configuration list}
}
\description{
Merge katex and user dependencies into package.json
}
\keyword{internal}
