% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{normalize_local_link}
\alias{normalize_local_link}
\title{Normalize local links to lowercase for articles/reference paths}
\usage{
normalize_local_link(link, warn = TRUE)
}
\arguments{
\item{link}{Link string}

\item{warn}{Whether to warn when link is normalized}
}
\value{
Normalized link string
}
\description{
Normalize local links to lowercase for articles/reference paths
}
\keyword{internal}
