% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{parse_content_item}
\alias{parse_content_item}
\title{Parse a content item which can be a string or a list with slug/label}
\usage{
parse_content_item(content)
}
\arguments{
\item{content}{Content item (string or list)}
}
\value{
List with slug and label
}
\description{
Parse a content item which can be a string or a list with slug/label
}
\keyword{internal}
