% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_md.R
\name{parse_format_lines}
\alias{parse_format_lines}
\title{Parse format lines into column definitions}
\usage{
parse_format_lines(format_lines)
}
\arguments{
\item{format_lines}{vector of lines that look like column definitions}
}
\value{
data frame with name and description columns
}
\description{
Parse format lines into column definitions
}
