% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{parse_namespace_exports}
\alias{parse_namespace_exports}
\title{Parse NAMESPACE file to extract exported functions}
\usage{
parse_namespace_exports(namespace_path)
}
\arguments{
\item{namespace_path}{Path to NAMESPACE file}
}
\value{
Character vector of exported function names
}
\description{
Parse NAMESPACE file to extract exported functions
}
\keyword{internal}
