% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preview_site}
\alias{preview_site}
\title{Preview the generated site}
\usage{
preview_site(output_path)
}
\arguments{
\item{output_path}{Path to built site}
}
\description{
Preview the generated site
}
\keyword{internal}
