% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_arguments.R
\name{process_arguments}
\alias{process_arguments}
\title{Formats function arguments into markdown table}
\usage{
process_arguments(arguments_text)
}
\arguments{
\item{arguments_text}{text of arguments from rd_obj}
}
\value{
dataframe of arguments tables with name and description columns
}
\description{
Formats function arguments into markdown table
}
\examples{
\dontrun{
rd_files <- extract_package_rd_contents("rdstarlight")
process_arguments(rd_files[['write_md_files.Rd']]$arguments)
}
}
