% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{process_articles}
\alias{process_articles}
\title{Process articles from vignettes/*.Rmd}
\usage{
process_articles(pkg_path, output_path, config)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Process articles from vignettes/*.Rmd
}
