% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_articles_and_readme}
\alias{process_articles_and_readme}
\title{Process articles (vignettes) and README together}
\usage{
process_articles_and_readme(pkg_path, output_path, config)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Builds all Rmd files in a single devtools::build_rmd() call to avoid
multiple package installs. Processes vignettes from config and README.Rmd.
"readme" in the config is treated specially - maps to README.Rmd at package root.
}
\keyword{internal}
