% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_news}
\alias{process_news}
\title{Process NEWS.md file}
\usage{
process_news(pkg_path, output_path, config)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Process NEWS.md file
}
\keyword{internal}
