% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{process_package_documentation}
\alias{process_package_documentation}
\title{Process package documentation into Starlight format}
\usage{
process_package_documentation(pkg_path, output_path, config)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Process package documentation into Starlight format
}
