% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{process_rmd_file}
\alias{process_rmd_file}
\title{Process a single .Rmd file to Markdown}
\usage{
process_rmd_file(rmd_path, output_dir)
}
\arguments{
\item{rmd_path}{Path to .Rmd file}

\item{output_dir}{Output directory for markdown file}
}
\description{
Process a single .Rmd file to Markdown
}
