% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{rd_contains_function}
\alias{rd_contains_function}
\title{Check if an Rd object contains documentation for a specific function}
\usage{
rd_contains_function(rd_obj, function_name)
}
\arguments{
\item{rd_obj}{Rd object from tools::Rd_db}

\item{function_name}{Function name to look for}
}
\value{
Logical indicating if the function is documented in this Rd object
}
\description{
Check if an Rd object contains documentation for a specific function
}
\keyword{internal}
