% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read and parse starlightr configuration}
\usage{
read_config(config_path)
}
\arguments{
\item{config_path}{Path to the _starlightr.yaml file}
}
\value{
List containing parsed configuration
}
\description{
Read and parse starlightr configuration
}
\keyword{internal}
