% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{read_config_toml}
\alias{read_config_toml}
\title{Read and parse TOML config file}
\usage{
read_config_toml(config_path = "_starlightr.toml")
}
\arguments{
\item{config_path}{Path to config file}
}
\value{
Parsed config as list
}
\description{
Read and parse TOML config file
}
\keyword{internal}
