% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_postprocess.R
\name{remove_sections}
\alias{remove_sections}
\title{Remove unwanted sections from Markdown}
\usage{
remove_sections(md, skip_sections)
}
\arguments{
\item{md}{Markdown string}

\item{skip_sections}{Character vector of section names to remove}
}
\value{
Markdown string with sections removed
}
\description{
Remove unwanted sections from Markdown
}
\keyword{internal}
