% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{require_pandoc}
\alias{require_pandoc}
\title{Ensure pandoc is available, error if not}
\usage{
require_pandoc()
}
\value{
NULL (called for side effect)
}
\description{
Ensure pandoc is available, error if not
}
\keyword{internal}
