% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{resolve_config_rmd_files}
\alias{resolve_config_rmd_files}
\title{Resolve config sidebar.articles to .Rmd file paths}
\usage{
resolve_config_rmd_files(pkg_path, config)
}
\arguments{
\item{pkg_path}{Package directory path}

\item{config}{Parsed config list}
}
\value{
Character vector of .Rmd file paths
}
\description{
If config has sidebar.articles, resolves slugs to .Rmd files.
Otherwise discovers all vignettes + README.
}
\keyword{internal}
