% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_convert.R
\name{section_has_equations}
\alias{section_has_equations}
\title{Check if an Rd section contains equations}
\usage{
section_has_equations(rd_section)
}
\arguments{
\item{rd_section}{An Rd section}
}
\value{
Logical indicating if equations are present
}
\description{
Check if an Rd section contains equations
}
\keyword{internal}
