% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{setup_starlight_structure}
\alias{setup_starlight_structure}
\title{Set up Starlight directory structure}
\usage{
setup_starlight_structure(output_path, config)
}
\arguments{
\item{output_path}{Path to output directory}

\item{config}{Configuration list}
}
\description{
Set up Starlight directory structure
}
\keyword{internal}
