% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{update_index_hero_image}
\alias{update_index_hero_image}
\title{Update hero image in existing index.mdx}
\usage{
update_index_hero_image(index_path, config)
}
\arguments{
\item{index_path}{Path to index.mdx file}

\item{config}{Configuration list}
}
\description{
Update hero image in existing index.mdx
}
