% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_starlightr.R
\name{update_rbuildignore}
\alias{update_rbuildignore}
\title{Update .Rbuildignore to include _starlightr.toml}
\usage{
update_rbuildignore(pkg_path)
}
\arguments{
\item{pkg_path}{Path to package directory}
}
\description{
Update .Rbuildignore to include _starlightr.toml
}
\keyword{internal}
