% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{validate_link_target}
\alias{validate_link_target}
\title{Validate that an internal link points to existing content}
\usage{
validate_link_target(link, context, pkg_path, exported)
}
\arguments{
\item{link}{Link string}

\item{context}{Context string for error messages}

\item{pkg_path}{Path to package directory}

\item{exported}{Character vector of exported function names}
}
\value{
TRUE if an issue was found, FALSE otherwise
}
\description{
Validate that an internal link points to existing content
}
\keyword{internal}
