% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{validate_version_config}
\alias{validate_version_config}
\title{Validate version configuration}
\usage{
validate_version_config(config)
}
\arguments{
\item{config}{Configuration list}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate version configuration
}
\keyword{internal}
