% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{write_config_toml}
\alias{write_config_toml}
\title{Write config list back to TOML file}
\usage{
write_config_toml(config, config_path = "_starlightr.toml")
}
\arguments{
\item{config}{Config list to write}

\item{config_path}{Path to config file}
}
\description{
Write config list back to TOML file
}
\keyword{internal}
