% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_md_files.R
\name{write_md_files}
\alias{write_md_files}
\title{Writes the parsed and formatted Rd file to a new file in output_dir}
\usage{
write_md_files(
  rd_files,
  output_dir,
  file_ext = ".mdx",
  code_sections = c("usage", "examples"),
  skip_sections = c("name", "alias", "title", "seealso")
)
}
\arguments{
\item{rd_files}{list of rd_objects to create md file for}

\item{output_dir}{path to directory to save new md files}

\item{file_ext}{either .md or .mdx. extension to use for files, default .md}

\item{code_sections}{character vector of sections to format as code blocks}

\item{skip_sections}{character vector of sections to skip entirely}
}
\description{
Writes the parsed and formatted Rd file to a new file in output_dir
}
\examples{
\dontrun{
rd_files <- extract_package_rd_content("starlightr")
output_dir <- file.path("path/to/starlightr/docs/content/rd_files")

write_md_files(rd_files, output_dir)

# Custom sections
write_md_files(
  rd_files,
  output_dir,
  sections = c("description", "usage", "examples")
)
}
}
