% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{create_index_page}
\alias{create_index_page}
\title{Create default index.mdx page}
\usage{
create_index_page(pkg_path, output_path, config, overwrite = FALSE)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}

\item{overwrite}{Logical, whether to overwrite existing index.mdx}
}
\description{
Create default index.mdx page
}
\keyword{internal}
