% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_package_rd_content}
\alias{extract_package_rd_content}
\title{extracts all Rd files from a package and retrieves each section
from extract_rd_section}
\usage{
extract_package_rd_content(pkg_name)
}
\arguments{
\item{pkg_name}{name of package to collect Rd objects for}
}
\value{
list of rd_objects, with each section parsed
}
\description{
extracts all Rd files from a package and retrieves each section
from extract_rd_section
}
\examples{
\dontrun{
extract_package_rd_content("rdstarlight")
}
}
