% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_rd_content}
\alias{extract_rd_content}
\title{Wraps extract_section for name, description, usage, arguments, value, examples}
\usage{
extract_rd_content(rd_obj)
}
\arguments{
\item{rd_obj}{rd object from tools::Rd_db}
}
\value{
list of sections for the rd_obj
}
\description{
Wraps extract_section for name, description, usage, arguments, value, examples
}
\examples{
\dontrun{
rd_files <- tools::Rd_db("rdstarlight")

extract_rd_content(rd_files[['extract_section.Rd']])
}
}
