% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_md.R
\name{format_md}
\alias{format_md}
\title{Formats the rd file contents to markdown}
\usage{
format_md(content)
}
\arguments{
\item{content}{rd_file contents}
}
\value{
markdown formatted string of content
}
\description{
Formats the rd file contents to markdown
}
\examples{
\dontrun{
rd_files <- extract_package_rd_contents("rdstarlight")
format_md(rd_files[["function_name.Rd"]])
}
}
