% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_custom_css}
\alias{generate_custom_css}
\title{Generate custom.css file for Starlight site}
\usage{
generate_custom_css(output_path)
}
\arguments{
\item{output_path}{Path to output directory}
}
\description{
Generate custom.css file for Starlight site
}
\keyword{internal}
