% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{generate_deploy_workflow}
\alias{generate_deploy_workflow}
\title{Generate deploy-docs.yml GitHub Actions workflow}
\usage{
generate_deploy_workflow(output_path, config)
}
\arguments{
\item{output_path}{Path to docs output directory}

\item{config}{Configuration list}
}
\description{
Generate deploy-docs.yml GitHub Actions workflow
}
\keyword{internal}
