% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_package_documentation}
\alias{process_package_documentation}
\title{Process package documentation into Starlight format}
\usage{
process_package_documentation(pkg_path, output_path, config, verbose = FALSE)
}
\arguments{
\item{pkg_path}{Path to package directory}

\item{output_path}{Path to output directory}

\item{config}{Configuration list}

\item{verbose}{Logical, whether to print debug messages for example capture}
}
\description{
Process package documentation into Starlight format
}
\keyword{internal}
